Update URLs to use the .org tld
The objective-see.com URLs are redirected to their objective-see.org
counterparts. An HTTP requests to the `PRODUCT_URL` will therefore
result in a `301` response and a subsequent request to the URL from the
`location` header of the response.

Update all URLs to skip that step and request the correct URL in the
first attempt.